Split out UserProfile
